Remove jobs from namespace
