+++ app/jobs/headlines/scheduled/daily.rb
@@ -1,0 +1,11 @@
+module Jobs
+  module Headlines
+    class Daily < Jobs::Headlines::Base
+      every 1.day
+
+      def initialize
+        @notification_type = "daily"
+      end
+    end
+  end
+end
