--- app/jobs/headlines/scheduled/scan_favorite_domains/daily.rb
@@ -1,13 +1,0 @@
-module Jobs
-  module Headlines
-    module ScanFavoriteDomains
-      class Daily < Jobs::Headlines::ScanFavoriteDomains::Base
-        every 1.day
-
-        def initialize
-          @notification_type = "daily"
-        end
-      end
-    end
-  end
-end
