--- app/jobs/headlines/scheduled/scan_favorite_domains/monthly.rb
@@ -1,13 +1,0 @@
-module Jobs
-  module Headlines
-    module ScanFavoriteDomains
-      class Monthly < Jobs::Headlines::ScanFavoriteDomains::Base
-        every 1.month
-
-        def initialize
-          @notification_type = "monthly"
-        end
-      end
-    end
-  end
-end
