--- app/jobs/headlines/scheduled/scan_favorite_domains/weekly.rb
@@ -1,13 +1,0 @@
-module Jobs
-  module Headlines
-    module ScanFavoriteDomains
-      class Weekly < Jobs::Headlines::ScanFavoriteDomains::Base
-        every 1.week
-
-        def initialize
-          @notification_type = "weekly"
-        end
-      end
-    end
-  end
-end
